fix(meaningful-comments): Refactored typewriter function to new module, formatted code, and added in meaningful comments
