+++ static/components/navbar.js
@@ -1,4 +1,7 @@
-import { isMobile, addComponentStyleSheetToHead } from "../scripts/utils/index.js";
+import {
+  isMobile,
+  addComponentStyleSheetToHead,
+} from "../scripts/utils/index.js";
 
 // Param is the navbar element to put in the children
 const navbar = (navbarSelector) => {
