+++ static/package.json
@@ -7,6 +7,6 @@
     "format": "find . \\( -name \"*.js\" -o -name \"*.html\" -o -name \"*.css\" \\) -not -path \"./node_modules/*\" | grep -v \"node_modules\" | xargs npx prettier --write"
   },
   "devDependencies": {
-    "prettier": "^3.2.5",
+    "prettier": "^3.2.5"
   }
 }
