+++ static/scripts/utils/index.js
@@ -1,7 +1,14 @@
 import addComponentStyleSheetToHead from "./addComponentStyleSheetToHead.js";
 import addHeadContent from "./addHeadContent.js";
 import setGlobalStyles from "./setGlobalStyles.js";
-import { isMobile, isMediumWidth, isTabletWidth, isDesktop } from "./other.js";
+import {
+  isMobile,
+  isMediumWidth,
+  isTabletWidth,
+  isDesktop,
+} from "./screenSizes.js";
+import { sleep } from "./other.js";
+import typewriter from "./animations/typewriter.js";
 
 export {
   addComponentStyleSheetToHead,
@@ -11,4 +18,6 @@ export {
   isMediumWidth,
   isTabletWidth,
   isDesktop,
+  typewriter,
+  sleep,
 };
