+++ static/scripts/utils/other.js
@@ -1,21 +1,6 @@
-const mobileWidth = 450;
-const tabletWidth = 750;
-const desktopWidth = 1000;
-
-const isMobile = () => {
-  return window.innerWidth <= mobileWidth;
+// Helper function to simulate sleep behavior, by returning a promise.
+const sleep = (ms) => {
+  return new Promise((resolve) => setTimeout(resolve, ms));
 };
 
-const isDesktop = () => {
-  return window.innerWidth >= 1000;
-};
-
-const isMediumWidth = () => {
-  return window.innerWidth > mobileWidth && window.innerWidth < tabletWidth;
-};
-
-const isTabletWidth = () => {
-  return window.innerWidth >= tabletWidth && window.innerWidth < desktopWidth;
-};
-
-export { isMobile, isMediumWidth, isTabletWidth, isDesktop };
+export { sleep };
